feat(MainMethodChecker): support new main method formats (#18)
New Java versions supports main methods without public and static
modifiers. We want the main method finder to find such methods, too.

In Java 23, this is still a previous feature. However, we want to
prepare this already since it seems likely that this feature will
mature from preview to released in Java 24.

Note: When using on an older JDK before this feature is fully released,
the main method finder may report false-positives.

Closes: #3